1/new structure - Suite (EndToEnd) 2/outlook 3/committed ui tests
