+++ pom.xml
@@ -136,7 +136,7 @@
                     <version>2.20</version>
                     <configuration>
                         <parallel>classes</parallel>
-                        <threadCount>2</threadCount>
+                        <threadCount>3</threadCount>
                         <argLine>
                             -javaagent:"${settings.localRepository}/org/aspectj/aspectjweaver/${aspectj.version}/aspectjweaver-${aspectj.version}.jar"
                         </argLine>
