+++ src/test/java/outlookTest/AutorizationOutlook.java
@@ -19,7 +19,7 @@ public class AutorizationOutlook extends ParentTest {
         outlookPage.clickOnButtonOnOnAutorizationForm();
         outlookPage.clickOnButtonOkQuiteFromSystem();
         outlookPage.chooseCurrentMail();
-        outlookPage.openLink();
+      //  outlookPage.openLink();
     }
